+++ app/src/main/java/com/dsm/gym/presentation/ui/fragment/sign/login/SignInFragment.kt
@@ -1,4 +1,4 @@
-package com.dsm.gym.presentation.ui.fragment
+package com.dsm.gym.presentation.ui.fragment.sign.login
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.dsm.gym.R
 
-class SecondRegisterFragment : Fragment() {
+class SignInFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_second_register, container, false)
+        return inflater.inflate(R.layout.fragment_login, container, false)
     }
 }
