+++ app/src/main/res/layout/fragment_complete_register.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.fragment.CompleteRegisterFragment">
+    tools:context=".presentation.ui.fragment.sign.register.complete.CompleteRegisterFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/register_complete_tv"
