+++ app/src/main/res/layout/fragment_first_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorDefault"
-    tools:context=".presentation.ui.fragment.FirstRegisterFragment">
+    tools:context=".presentation.ui.fragment.sign.register.FirstRegisterFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:layout_width="76dp"
