+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorDefault"
-    tools:context=".presentation.ui.fragment.SignInFragment">
+    tools:context=".presentation.ui.fragment.sign.login.SignInFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:layout_width="76dp"
